feat: 프로젝트 로그인 오류 및 캐러셀 추가
