+++ step-2-week1/.env
@@ -1,0 +1,5 @@
+REACT_APP_API_URL="http://kakao-app-env.eba-kfsgeb74.ap-northeast-2.elasticbeanstalk.com"
+REACT_APP_API_TITLE=카카오쇼핑
+REACT_APP_PORT=3000
+REACT_APP_API_TIMEOUT=1000
+GENERATE_SOURCEMAP=false
