+++ step-2-week1/.gitignore
@@ -14,9 +14,6 @@
 /build
 
 
-package.lock.json
-
-.package.lock.json
 
 # misc
 .DS_Store
