+++ step-2-week1/default.conf
@@ -1,0 +1,11 @@
+server {
+    server_name _;
+
+    location / {
+        proxy-pass http://localhost:3000;
+    }
+
+    location /api/ {
+        proxy_pass http://backend-service.default.svc.cluster.local:8080/;
+    }
+}
