+++ step-2-week1/package.json
@@ -1,6 +1,6 @@
 {
   "name": "step-2-wee1",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.5",
