+++ step-2-week1/src/pages/HomePage.jsx
@@ -2,6 +2,8 @@ import ProductGrid from "../component/organisms/ProductGrid";
 import { useQuery } from "react-query";
 import { fetchProducts } from "../services/product";
 import Loader from "../component/atoms/Loader";
+import Carousel from "../component/atoms/Carousel";
+import React from 'react';
 
 const HomePage = () => {
   const { data: products, isLoading } = useQuery(["products"], fetchProducts, {
@@ -12,9 +14,13 @@ const HomePage = () => {
 
 
   return (
+
     <div>
+      <Carousel /> 
       <h1>Home Page</h1>
         {isLoading ? ( <Loader /> ) : ( <ProductGrid productsData={products} /> )}
+
+        
     </div> 
     );
