+++ step-2-week1/src/services/index.js
@@ -4,7 +4,7 @@ import axios from "axios";
 
 export const instance = axios.create({
   baseURL: process.env.REACT_APP_API_URL,
-  timeout: 1000,
+  timeout: process.env.REACT_APP_TIMEOUT,
   headers: {
     "Content-Type": "application/json",
     // "Authorization": `Bearer ${localStorage.getItem('token')}`,
